+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/grammar/MetricAlertConditionParser.py
@@ -15,7 +15,7 @@ import sys
 def serializedATN():
     with StringIO() as buf:
         buf.write(u"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3")
-        buf.write(u"\17{\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7")
+        buf.write(u"\20{\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7")
         buf.write(u"\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r\4")
         buf.write(u"\16\t\16\4\17\t\17\4\20\t\20\3\2\3\2\3\2\3\2\7\2%\n\2")
         buf.write(u"\f\2\16\2(\13\2\3\2\3\2\3\2\3\2\3\2\3\2\5\2\60\n\2\3")
@@ -26,32 +26,32 @@ def serializedATN():
         buf.write(u"\13\3\13\3\f\3\f\3\f\3\r\3\r\3\r\3\16\3\16\3\16\3\17")
         buf.write(u"\3\17\3\17\3\17\7\17q\n\17\f\17\16\17t\13\17\3\20\6\20")
         buf.write(u"w\n\20\r\20\16\20x\3\20\2\2\21\2\4\6\b\n\f\16\20\22\24")
-        buf.write(u"\26\30\32\34\36\2\7\4\2\r\r\17\17\4\2\4\4\6\6\3\2\7\b")
-        buf.write(u"\4\2\4\4\t\t\5\2\13\13\r\r\17\17\2s\2 \3\2\2\2\4@\3\2")
-        buf.write(u"\2\2\6C\3\2\2\2\bF\3\2\2\2\nJ\3\2\2\2\fM\3\2\2\2\16O")
-        buf.write(u"\3\2\2\2\20R\3\2\2\2\22\\\3\2\2\2\24`\3\2\2\2\26c\3\2")
-        buf.write(u"\2\2\30f\3\2\2\2\32i\3\2\2\2\34l\3\2\2\2\36v\3\2\2\2")
-        buf.write(u" &\5\4\3\2!\"\5\6\4\2\"#\7\3\2\2#%\3\2\2\2$!\3\2\2\2")
-        buf.write(u"%(\3\2\2\2&$\3\2\2\2&\'\3\2\2\2\'/\3\2\2\2(&\3\2\2\2")
-        buf.write(u")*\7\f\2\2*+\5\b\5\2+,\7\f\2\2,-\7\r\2\2-\60\3\2\2\2")
-        buf.write(u".\60\5\b\5\2/)\3\2\2\2/.\3\2\2\2\60\61\3\2\2\2\61\62")
-        buf.write(u"\5\n\6\2\62\67\5\f\7\2\63\64\7\r\2\2\64\66\5\20\t\2\65")
-        buf.write(u"\63\3\2\2\2\669\3\2\2\2\67\65\3\2\2\2\678\3\2\2\28=\3")
-        buf.write(u"\2\2\29\67\3\2\2\2:<\7\16\2\2;:\3\2\2\2<?\3\2\2\2=;\3")
-        buf.write(u"\2\2\2=>\3\2\2\2>\3\3\2\2\2?=\3\2\2\2@A\7\17\2\2AB\7")
-        buf.write(u"\r\2\2B\5\3\2\2\2CD\7\17\2\2D\7\3\2\2\2EG\t\2\2\2FE\3")
-        buf.write(u"\2\2\2GH\3\2\2\2HF\3\2\2\2HI\3\2\2\2I\t\3\2\2\2JK\7\n")
-        buf.write(u"\2\2KL\7\r\2\2L\13\3\2\2\2MN\7\13\2\2N\r\3\2\2\2OP\7")
-        buf.write(u"\5\2\2PQ\7\r\2\2Q\17\3\2\2\2RS\5\16\b\2SY\5\22\n\2TU")
-        buf.write(u"\5\24\13\2UV\5\22\n\2VX\3\2\2\2WT\3\2\2\2X[\3\2\2\2Y")
-        buf.write(u"W\3\2\2\2YZ\3\2\2\2Z\21\3\2\2\2[Y\3\2\2\2\\]\5\32\16")
-        buf.write(u"\2]^\5\26\f\2^_\5\34\17\2_\23\3\2\2\2`a\t\3\2\2ab\7\r")
-        buf.write(u"\2\2b\25\3\2\2\2cd\t\4\2\2de\7\r\2\2e\27\3\2\2\2fg\t")
-        buf.write(u"\5\2\2gh\7\r\2\2h\31\3\2\2\2ij\7\17\2\2jk\7\r\2\2k\33")
-        buf.write(u"\3\2\2\2lr\5\36\20\2mn\5\30\r\2no\5\36\20\2oq\3\2\2\2")
-        buf.write(u"pm\3\2\2\2qt\3\2\2\2rp\3\2\2\2rs\3\2\2\2s\35\3\2\2\2")
-        buf.write(u"tr\3\2\2\2uw\t\6\2\2vu\3\2\2\2wx\3\2\2\2xv\3\2\2\2xy")
-        buf.write(u"\3\2\2\2y\37\3\2\2\2\n&/\67=HYrx")
+        buf.write(u"\26\30\32\34\36\2\7\5\2\3\4\16\16\20\20\4\2\5\5\7\7\3")
+        buf.write(u"\2\b\t\4\2\5\5\n\n\5\2\f\f\16\16\20\20\2s\2 \3\2\2\2")
+        buf.write(u"\4@\3\2\2\2\6C\3\2\2\2\bF\3\2\2\2\nJ\3\2\2\2\fM\3\2\2")
+        buf.write(u"\2\16O\3\2\2\2\20R\3\2\2\2\22\\\3\2\2\2\24`\3\2\2\2\26")
+        buf.write(u"c\3\2\2\2\30f\3\2\2\2\32i\3\2\2\2\34l\3\2\2\2\36v\3\2")
+        buf.write(u"\2\2 &\5\4\3\2!\"\5\6\4\2\"#\7\3\2\2#%\3\2\2\2$!\3\2")
+        buf.write(u"\2\2%(\3\2\2\2&$\3\2\2\2&\'\3\2\2\2\'/\3\2\2\2(&\3\2")
+        buf.write(u"\2\2)*\7\r\2\2*+\5\b\5\2+,\7\r\2\2,-\7\16\2\2-\60\3\2")
+        buf.write(u"\2\2.\60\5\b\5\2/)\3\2\2\2/.\3\2\2\2\60\61\3\2\2\2\61")
+        buf.write(u"\62\5\n\6\2\62\67\5\f\7\2\63\64\7\16\2\2\64\66\5\20\t")
+        buf.write(u"\2\65\63\3\2\2\2\669\3\2\2\2\67\65\3\2\2\2\678\3\2\2")
+        buf.write(u"\28=\3\2\2\29\67\3\2\2\2:<\7\17\2\2;:\3\2\2\2<?\3\2\2")
+        buf.write(u"\2=;\3\2\2\2=>\3\2\2\2>\3\3\2\2\2?=\3\2\2\2@A\7\20\2")
+        buf.write(u"\2AB\7\16\2\2B\5\3\2\2\2CD\7\20\2\2D\7\3\2\2\2EG\t\2")
+        buf.write(u"\2\2FE\3\2\2\2GH\3\2\2\2HF\3\2\2\2HI\3\2\2\2I\t\3\2\2")
+        buf.write(u"\2JK\7\13\2\2KL\7\16\2\2L\13\3\2\2\2MN\7\f\2\2N\r\3\2")
+        buf.write(u"\2\2OP\7\6\2\2PQ\7\16\2\2Q\17\3\2\2\2RS\5\16\b\2SY\5")
+        buf.write(u"\22\n\2TU\5\24\13\2UV\5\22\n\2VX\3\2\2\2WT\3\2\2\2X[")
+        buf.write(u"\3\2\2\2YW\3\2\2\2YZ\3\2\2\2Z\21\3\2\2\2[Y\3\2\2\2\\")
+        buf.write(u"]\5\32\16\2]^\5\26\f\2^_\5\34\17\2_\23\3\2\2\2`a\t\3")
+        buf.write(u"\2\2ab\7\16\2\2b\25\3\2\2\2cd\t\4\2\2de\7\16\2\2e\27")
+        buf.write(u"\3\2\2\2fg\t\5\2\2gh\7\16\2\2h\31\3\2\2\2ij\7\20\2\2")
+        buf.write(u"jk\7\16\2\2k\33\3\2\2\2lr\5\36\20\2mn\5\30\r\2no\5\36")
+        buf.write(u"\20\2oq\3\2\2\2pm\3\2\2\2qt\3\2\2\2rp\3\2\2\2rs\3\2\2")
+        buf.write(u"\2s\35\3\2\2\2tr\3\2\2\2uw\t\6\2\2vu\3\2\2\2wx\3\2\2")
+        buf.write(u"\2xv\3\2\2\2xy\3\2\2\2y\37\3\2\2\2\n&/\67=HYrx")
         return buf.getvalue()
 
 
@@ -65,11 +65,12 @@ class MetricAlertConditionParser ( Parser ):
 
     sharedContextCache = PredictionContextCache()
 
-    literalNames = [ u"<INVALID>", u"'.'", u"','" ]
+    literalNames = [ u"<INVALID>", u"'.'", u"'/'", u"','" ]
 
-    symbolicNames = [ u"<INVALID>", u"<INVALID>", u"<INVALID>", u"WHERE", 
-                      u"AND", u"INCLUDES", u"EXCLUDES", u"OR", u"OPERATOR", 
-                      u"NUMBER", u"QUOTE", u"WHITESPACE", u"NEWLINE", u"WORD" ]
+    symbolicNames = [ u"<INVALID>", u"<INVALID>", u"<INVALID>", u"<INVALID>", 
+                      u"WHERE", u"AND", u"INCLUDES", u"EXCLUDES", u"OR", 
+                      u"OPERATOR", u"NUMBER", u"QUOTE", u"WHITESPACE", u"NEWLINE", 
+                      u"WORD" ]
 
     RULE_expression = 0
     RULE_aggregation = 1
@@ -95,17 +96,18 @@ class MetricAlertConditionParser ( Parser ):
     EOF = Token.EOF
     T__0=1
     T__1=2
-    WHERE=3
-    AND=4
-    INCLUDES=5
-    EXCLUDES=6
-    OR=7
-    OPERATOR=8
-    NUMBER=9
-    QUOTE=10
-    WHITESPACE=11
-    NEWLINE=12
-    WORD=13
+    T__2=3
+    WHERE=4
+    AND=5
+    INCLUDES=6
+    EXCLUDES=7
+    OR=8
+    OPERATOR=9
+    NUMBER=10
+    QUOTE=11
+    WHITESPACE=12
+    NEWLINE=13
+    WORD=14
 
     def __init__(self, input, output=sys.stdout):
         super(MetricAlertConditionParser, self).__init__(input, output=output)
@@ -218,7 +220,7 @@ class MetricAlertConditionParser ( Parser ):
                 self.state = 42
                 self.match(MetricAlertConditionParser.WHITESPACE)
                 pass
-            elif token in [MetricAlertConditionParser.WHITESPACE, MetricAlertConditionParser.WORD]:
+            elif token in [MetricAlertConditionParser.T__0, MetricAlertConditionParser.T__1, MetricAlertConditionParser.WHITESPACE, MetricAlertConditionParser.WORD]:
                 self.state = 44
                 self.metric()
                 pass
@@ -387,7 +389,7 @@ class MetricAlertConditionParser ( Parser ):
             while True:
                 self.state = 67
                 _la = self._input.LA(1)
-                if not(_la==MetricAlertConditionParser.WHITESPACE or _la==MetricAlertConditionParser.WORD):
+                if not((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << MetricAlertConditionParser.T__0) | (1 << MetricAlertConditionParser.T__1) | (1 << MetricAlertConditionParser.WHITESPACE) | (1 << MetricAlertConditionParser.WORD))) != 0)):
                     self._errHandler.recoverInline(self)
                 else:
                     self._errHandler.reportMatch(self)
@@ -395,7 +397,7 @@ class MetricAlertConditionParser ( Parser ):
                 self.state = 70 
                 self._errHandler.sync(self)
                 _la = self._input.LA(1)
-                if not (_la==MetricAlertConditionParser.WHITESPACE or _la==MetricAlertConditionParser.WORD):
+                if not ((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << MetricAlertConditionParser.T__0) | (1 << MetricAlertConditionParser.T__1) | (1 << MetricAlertConditionParser.WHITESPACE) | (1 << MetricAlertConditionParser.WORD))) != 0)):
                     break
 
         except RecognitionException as re:
@@ -585,7 +587,7 @@ class MetricAlertConditionParser ( Parser ):
             self.state = 87
             self._errHandler.sync(self)
             _la = self._input.LA(1)
-            while _la==MetricAlertConditionParser.T__1 or _la==MetricAlertConditionParser.AND:
+            while _la==MetricAlertConditionParser.T__2 or _la==MetricAlertConditionParser.AND:
                 self.state = 82
                 self.dim_separator()
                 self.state = 83
@@ -689,7 +691,7 @@ class MetricAlertConditionParser ( Parser ):
             self.enterOuterAlt(localctx, 1)
             self.state = 94
             _la = self._input.LA(1)
-            if not(_la==MetricAlertConditionParser.T__1 or _la==MetricAlertConditionParser.AND):
+            if not(_la==MetricAlertConditionParser.T__2 or _la==MetricAlertConditionParser.AND):
                 self._errHandler.recoverInline(self)
             else:
                 self._errHandler.reportMatch(self)
@@ -792,7 +794,7 @@ class MetricAlertConditionParser ( Parser ):
             self.enterOuterAlt(localctx, 1)
             self.state = 100
             _la = self._input.LA(1)
-            if not(_la==MetricAlertConditionParser.T__1 or _la==MetricAlertConditionParser.OR):
+            if not(_la==MetricAlertConditionParser.T__2 or _la==MetricAlertConditionParser.OR):
                 self._errHandler.recoverInline(self)
             else:
                 self._errHandler.reportMatch(self)