+++ src/command_modules/azure-cli-rdbms/azure/cli/command_modules/rdbms/_help.py
@@ -26,10 +26,10 @@ def add_helps(command_group, server_type):
                       text: az {1} server create -l northeurope -g testgroup -n testsvr -u username -p password
                     - name: Create a {0} server with a Standard performance tier and 2 vcore in North Europe.
                       text: az {1} server create -l northeurope -g testgroup -n testsvr -u username -p password \\
-                            --sku-name GP_Gen4_2
+                            --sku-name GP_Gen5_2
                     - name: Create a {0} server with all paramaters set.
                       text: az {1} server create -l northeurope -g testgroup -n testsvr -u username -p password \\
-                            --sku-name B_Gen4_2 --ssl-enforcement Disabled \\
+                            --sku-name B_Gen5_2 --ssl-enforcement Disabled \\
                             --storage-size 51200 --tags "key=value" --version {{server-version}}
                 """.format(server_type, command_group)
     helps['{} server restore'.format(command_group)] = """
@@ -61,7 +61,7 @@ def add_helps(command_group, server_type):
                 short-summary: Update a server.
                 examples:
                     - name: Update a server's sku.
-                      text: az {0} server update -g testgroup -n testsvrnew --sku-name GP_Gen4_4
+                      text: az {0} server update -g testgroup -n testsvrnew --sku-name GP_Gen5_4
                     - name: Update a server's tags.
                       text: az {0} server update -g testgroup -n testsvrnew --tags "k1=v1" "k2=v2"
                 """.format(command_group)