+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/tests/latest/test_resource.py
@@ -155,6 +155,53 @@ class ResourceIDScenarioTest(ScenarioTest):
         self.cmd('resource delete --id {vnet_id}', checks=self.is_empty())
 
 
+class ResourceGenericUpdate(LiveScenarioTest):
+    @ResourceGroupPreparer(name_prefix='cli_test_resource_id')
+    def test_resource_id_scenario(self, resource_group):
+        self.kwargs.update({
+            'stor_1': self.create_random_name(prefix='stor1', length=10),
+            'stor_2': self.create_random_name(prefix='stor2', length=10)
+        })
+
+        # create storage accounts
+        self.cmd('az storage account create -g {rg} -n {stor_1}')
+        self.cmd('az storage account create -g {rg} -n {stor_2}')
+
+        # get ids
+        self.kwargs['stor_ids'] = " ".join(self.cmd('az storage account list -g {rg} --query "[].id"').get_output_in_json())
+
+        # update tags
+        self.cmd('az storage account update --ids {stor_ids} --set tags.isTag=True tags.isNotTag=False')
+
+        self.cmd('az storage account show --name {stor_1} -g {rg}', checks=[
+            self.check('tags.isTag', 'True'),
+            self.check('tags.isNotTag', 'False')
+        ])
+        self.cmd('az storage account show --name {stor_2} -g {rg}', checks=[
+            self.check('tags.isTag', 'True'),
+            self.check('tags.isNotTag', 'False')
+        ])
+
+        # delete tags.isTag
+        self.cmd('az storage account update --ids {stor_ids} --remove tags.isTag')
+
+        self.cmd('az storage account show --name {stor_1} -g {rg} --query "tags"', checks=[
+            self.check('isNotTag', 'False'),
+            self.check('isTag', None)
+        ])
+        self.cmd('az storage account show --name {stor_2} -g {rg} --query "tags"', checks=[
+            self.check('isNotTag', 'False'),
+            self.check('isTag', None)
+        ])
+
+        # delete tags.isNotTag
+        self.cmd('az storage account update --ids {stor_ids} --remove tags.isNotTag')
+
+        # check tags is empty.
+        self.cmd('az storage account show --name {stor_1} -g {rg} --query "tags"', checks=self.is_empty())
+        self.cmd('az storage account show --name {stor_2} -g {rg} --query "tags"', checks=self.is_empty())
+
+
 class ResourceCreateAndShowScenarioTest(ScenarioTest):
 
     @ResourceGroupPreparer(name_prefix='cli_test_resource_create')
@@ -879,6 +926,18 @@ class InvokeActionTest(ScenarioTest):
         self.kwargs['request_body'] = '{\\"vhdPrefix\\":\\"myPrefix\\",\\"destinationContainerName\\":\\"container\\",\\"overwriteVhds\\":\\"true\\"}'
 
         self.cmd('resource invoke-action --action capture --ids {vm_id} --request-body {request_body}')
+
+
+class GlobalIdsScenarioTest(ScenarioTest):
+    @ResourceGroupPreparer(name_prefix='cli_test_global_ids')
+    def test_global_ids(self, resource_group):
+
+        self.kwargs.update({
+            'vnet': 'vnet1'
+        })
+        self.kwargs['vnet_id'] = self.cmd('network vnet create -g {rg} -n {vnet}').get_output_in_json()['newVNet']['id']
+        # command will fail if the other parameters were actually used
+        self.cmd('network vnet show --subscription fakesub --resource-group fakerg -n fakevnet --ids {vnet_id}')
 
 
 if __name__ == '__main__':