+++ src/azure-cli-core/HISTORY.rst
@@ -2,9 +2,19 @@
 
 Release History
 ===============
+
+2.0.50
+++++++
+* Fix issue where update commands using `--remove` and `--ids` fail after first update is applied to first resource in ids list.
+
+2.0.49
+++++++
+* Fix issue with `--ids` where `--subscription` would take precedence over the subscription in `--ids`.
+  Adding explicit warnings when name parameters would be ignored by use of `--ids`.
+
 2.0.48
 ++++++
-* Minor fixes
+* Fix Homebrew.
 
 2.0.47
 ++++++