+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+2.1.9
+++++++
+* `ad sp create-for-rbac`: clarify the confusion between displayName and service principal name
+* support grant permissions to AAD apps
+
+2.1.8
+++++++
+* support add/remove/list owner on AAD Applciation and Group objects
 
 2.1.7
 ++++++